requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aioredis==1.3.1
amqp==2.5.2
APScheduler==3.6.3
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.1
Automat==20.2.0
baidu-aip==2.2.18.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
billiard==3.6.3.0
brotlipy==0.7.0
cchardet==2.1.6
celery==4.4.1
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
cryptography==3.2
daphne==2.4.1
Django==3.0.7
django-autocomplete-light==3.5.1
django-cors-headers==3.2.1
django-filter==2.3.0
django-qiniu-storage==2.3.1
django-redis==4.11.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
dwebsocket==0.5.12
Fabric3==1.14.post1
gunicorn==20.0.4
h11==0.9.0
hiredis==1.0.1
httptools==0.1.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==17.5.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.1
kafka-python==2.0.1
kombu==4.6.8
lru-dict==1.1.6
lxml==4.5.0
MarkupSafe==1.1.1
msgpack==0.6.2
multidict==4.7.4
mysqlclient==1.4.6
oauthlib==3.1.0
packaging==20.4
paramiko==2.7.1
Pillow==7.0.0
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycparser==2.19
PyHamcrest==2.0.0
PyJWT==1.7.1
pymongo==3.10.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2019.3
qiniu==7.2.6
redis==3.4.1
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
supervisor==4.1.0
toml==0.10.2
Twisted==20.3.0
txaio==20.1.1
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.8
uvicorn==0.11.7
uvloop==0.14.0
vine==1.3.0
websockets==8.1
yarl==1.4.2
zipp==3.1.0
zope.interface==4.7.1
